Remove the dead remote sync machinery from SQLDatabase
Every shipped build has run SQLDatabase with the deleted object cache off.
DataContentProvider.onCreate builds the database and immediately calls
setDeletedObjectCacheEnabled(IS_REMOTE_SYNC_ENABLED), a private constant that
has read false since the initial commit, so the soft delete side of 37 branches
has never run. SyncContentProvider never calls the setter, so its own instance
ran with the constructor default of true, but it only issues raw table reads and
writes and never reaches a method one of those branches is in.

This deletes the flag, its setter and the constant, and keeps the branch each of
them selected: the hard delete side of the 37 if (mCacheDeletedObjects) branches,
and the body of the three if (!mCacheDeletedObjects) sites. Those three are not a
soft delete path. They sweep link rows that unconditional writes just above them
marked DELETED = 1, in updateTransactionRow, updateTransfer and updateDebtPeople.
The marker writes stay and the sweep now runs every time, which is what it
already did.

The deleted, uuid and last_edit columns stay and are still written, because the
backup format carries them. The 23 read filters on deleted = 0 stay as well, and
they are load bearing today. updateBudget flags every budget_wallet and
budget_category row of a budget deleted and then revives the ones still selected,
and unlike the three link tables above it has no sweep behind it, so a category
dropped from a budget leaves a deleted = 1 row sitting in budget_category for
good. Those filters are what keeps it out of the lists. None of the 23 is touched
here.

Separately, the five files under storage/database that imported picker.IconPicker
now import utils.IconLoader. All five wanted one static, getColorIconString, so
it moved to IconLoader with its body unchanged. The two call sites left inside
IconPicker, the one in TutorialActivity and the two lint baseline entries that
name the method were repointed to follow it.

SQLDatabase goes from 5779 lines to 5489.

Tested: 251 unit tests and the 91 SQLDatabaseTest cases green on this branch and
on master. deletePerson now also asks the person, debt_people,
transaction_people and transfer_people tables for rows still carrying the deleted
flag, because every accessor the suite goes through filters on deleted = 0 and so
cannot tell a removed row from a flagged one. Turning all five deletes in that
method into flag writes leaves it the only one of the 91 that fails. On the emulator, seven operations ran against a build of master and
again against this branch from a byte identical copy of the same database,
swapping the person on a transaction, deleting that person, the refusal to delete
a category that has children, deleting the child and then the parent, the refusal
to delete a wallet used in transfers, and deleting a wallet. All 25 tables match
row for row afterward, uuid and last_edit aside.
